Exception Handling Added

+++ src/main/java/com/neyesem/neyesembackend/service/RestaurantService.java
@@ -1,7 +1,11 @@
 package com.neyesem.neyesembackend.service;
 
 import com.neyesem.neyesembackend.entity.Restaurant;
+import com.neyesem.neyesembackend.exception.CustomException;
+import com.neyesem.neyesembackend.exception.EntityAlreadyExistsException;
+import com.neyesem.neyesembackend.exception.EntityNotFoundException;
 import com.neyesem.neyesembackend.repository.IRestaurantRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -22,37 +26,40 @@ public class RestaurantService {
         return restaurantRepository.findAll();
     }
 
-
     public void deleteRestaurantById(Long id){
-        restaurantRepository.deleteById(id);
+
+        if (restaurantRepository.existsById(id)) {
+            restaurantRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Restaurant",id);
+        }
+
     }
-
-    public void createRestaurant(Restaurant restaurant){
-        restaurantRepository.save(restaurant);
-    }
-
-
 
     public Restaurant getRestaurantById(Long id) {
-        return restaurantRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Restaurant not found with id: " + id));
+
+        Optional<Restaurant> restaurant = restaurantRepository.findById(id);
+
+        if (!restaurant.isPresent()){
+            throw new EntityNotFoundException("Restaurant",id);
+        }
+
+        return restaurant.get();
+
     }
 
     public Restaurant saveRestaurant(Restaurant newRestaurant){
         newRestaurant.setCreateDate(new Date(System.currentTimeMillis()));
-
         return restaurantRepository.save(newRestaurant);
     }
 
 
     public Restaurant updateRestaurantById(Long id, Restaurant updateRestaurant) {
 
-        Restaurant oldRestaurant = restaurantRepository.getRestaurantById(id);
+        Restaurant oldRestaurant = getRestaurantById(id);
         oldRestaurant.setAddress(updateRestaurant.getAddress());
         oldRestaurant.setGoogleMapsLink(updateRestaurant.getGoogleMapsLink());
-
-
-
         return restaurantRepository.save(oldRestaurant);
+
     }
 }
